requirements.txt
pdf2image
Pillow
streamlit
fastapi
python-multipart
uvicorn
